refactor(ops): consolidate horizontal scaling tests
